+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+- Fixed an error that could occur when editing element search field contents
+- Fixed a bug where previously selected elements for element search fields couldn't be removed
+
 ## 3.0.1 - 2022-10-19
 
 ### Fixed